tasks/extract-pwd-parcels/requirements.txt
google-cloud-storage>=2.10.0
functions-framework>=3.0.0
requests>=2.31.0
